consolidate logic into LiveData Kotlin file in preparation for adding multi-observeForever
