+++ gto-support-androidx-lifecycle/src/main/kotlin/org/ccci/gto/android/common/androidx/lifecycle/LiveData.kt
@@ -3,7 +3,45 @@ package org.ccci.gto.android.common.androidx.lifecycle
 import androidx.annotation.MainThread
 import androidx.lifecycle.LifecycleOwner
 import androidx.lifecycle.LiveData
+import androidx.lifecycle.MediatorLiveData
 import androidx.lifecycle.Observer
+
+// region Multi-observe
+fun <IN1, IN2> LifecycleOwner.observe(
+    source1: LiveData<IN1>,
+    source2: LiveData<IN2>,
+    observer: (IN1, IN2) -> Unit
+) = compoundLiveData(source1, source2).observe(this) {
+    @Suppress("UNCHECKED_CAST")
+    observer(source1.value as IN1, source2.value as IN2)
+}
+
+fun <IN1, IN2, IN3> LifecycleOwner.observe(
+    source1: LiveData<IN1>,
+    source2: LiveData<IN2>,
+    source3: LiveData<IN3>,
+    observer: (IN1, IN2, IN3) -> Unit
+) = compoundLiveData(source1, source2, source3).observe(this) {
+    @Suppress("UNCHECKED_CAST")
+    observer(source1.value as IN1, source2.value as IN2, source3.value as IN3)
+}
+
+private fun compoundLiveData(vararg input: LiveData<*>): LiveData<Unit> {
+    val result = MediatorLiveData<Unit>()
+    val state = object {
+        val inputInitialized = BooleanArray(input.size) { false }
+        var isInitialized = false
+            get() = field || inputInitialized.all { it }.also { field = it }
+    }
+    input.forEachIndexed { i, it ->
+        result.addSource(it) {
+            state.inputInitialized[i] = true
+            if (state.isInitialized) result.value = Unit
+        }
+    }
+    return result
+}
+// endregion Multi-observe
 
 // region observeOnce
 @MainThread
