+++ gto-support-androidx-lifecycle/src/test/kotlin/org/ccci/gto/android/common/androidx/lifecycle/LiveDataMultiObserveTest.kt
@@ -13,7 +13,7 @@ import org.mockito.kotlin.verify
 import org.mockito.kotlin.verifyNoMoreInteractions
 
 @OptIn(ExperimentalCoroutinesApi::class)
-class LifecycleOwnerMultiObserveTest : BaseLiveDataTest() {
+class LiveDataMultiObserveTest : BaseLiveDataTest() {
     private val lifecycleOwner = TestLifecycleOwner(Lifecycle.State.STARTED, UnconfinedTestDispatcher())
 
     private val data1 = MutableLiveData<String>()
@@ -21,7 +21,7 @@ class LifecycleOwnerMultiObserveTest : BaseLiveDataTest() {
     private val data3 = MutableLiveData<Long>()
 
     @Test
-    fun `Multi-observe - 2 LiveDatas`() {
+    fun `Multi-observe - LifecycleOwner - 2 LiveDatas`() {
         lifecycleOwner.observe(data1, data2) { d1, d2 -> observer.onChanged("$d1 $d2") }
         data1.value = "a"
         data2.value = 1
@@ -35,7 +35,7 @@ class LifecycleOwnerMultiObserveTest : BaseLiveDataTest() {
     }
 
     @Test
-    fun `Multi-observe - 3 LiveDatas`() {
+    fun `Multi-observe - LifecycleOwner - 3 LiveDatas`() {
         lifecycleOwner.observe(data1, data2, data3) { d1, d2, d3 -> observer.onChanged("$d1 $d2 $d3") }
         data1.value = "a"
         data2.value = 1
